+++ tests/hashtest.c
@@ -24,7 +24,7 @@ static virHashTablePtr
 testHashInit(int size)
 {
     virHashTablePtr hash;
-    int i;
+    ssize_t i;
 
     if (!(hash = virHashCreate(size, NULL)))
         return NULL;
@@ -33,14 +33,14 @@ testHashInit(int size)
      * collision list in the same order as in the uuids array
      */
     for (i = ARRAY_CARDINALITY(uuids) - 1; i >= 0; i--) {
-        int oldsize = virHashTableSize(hash);
+        ssize_t oldsize = virHashTableSize(hash);
         if (virHashAddEntry(hash, uuids[i], (void *) uuids[i]) < 0) {
             virHashFree(hash);
             return NULL;
         }
 
         if (virHashTableSize(hash) != oldsize && virTestGetDebug()) {
-            fprintf(stderr, "\nhash grown from %d to %d",
+            fprintf(stderr, "\nhash grown from %zd to %zd",
                     oldsize, virHashTableSize(hash));
         }
     }
@@ -70,19 +70,19 @@ testHashCheckForEachCount(void *payload ATTRIBUTE_UNUSED,
 }
 
 static int
-testHashCheckCount(virHashTablePtr hash, int count)
+testHashCheckCount(virHashTablePtr hash, size_t count)
 {
-    int iter_count = 0;
+    ssize_t iter_count = 0;
 
     if (virHashSize(hash) != count) {
-        testError("\nhash contains %d instead of %d elements\n",
+        testError("\nhash contains %zd instead of %zu elements\n",
                   virHashSize(hash), count);
         return -1;
     }
 
     iter_count = virHashForEach(hash, testHashCheckForEachCount, NULL);
     if (count != iter_count) {
-        testError("\nhash claims to have %d elements but iteration finds %d\n",
+        testError("\nhash claims to have %zu elements but iteration finds %zd\n",
                   count, iter_count);
         return -1;
     }
@@ -93,7 +93,7 @@ testHashCheckCount(virHashTablePtr hash, int count)
 
 struct testInfo {
     void *data;
-    int count;
+    size_t count;
 };
